pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-predictor-prod"
version = "0.1.0"
description = "Production-grade stock price prediction platform using LSTM and attention mechanisms"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ML Engineering Team", email = "team@example.com"}
]
keywords = ["stock-prediction", "lstm", "time-series", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    
    # Time Series Analysis
    "statsmodels>=0.14.0",
    "pmdarima>=2.0.3",
    "pandas-ta>=0.3.14b0",
    
    # Deep Learning
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    
    # Feature Engineering & ML
    "scikit-learn>=1.3.0",
    "polars>=0.19.0",
    
    # Data APIs
    "yfinance>=0.2.32",
    "alpha-vantage>=2.3.1",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    
    # Database & Caching
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    
    # ML Pipeline & Orchestration
    "airflow>=2.7.0",
    "mlflow>=2.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # API & Serving
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    
    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "python-json-logger>=2.0.7",
    "loguru>=0.7.0",
    
    # Data Validation
    "great-expectations>=0.17.0",
    "pandera>=0.17.0",
    
    # Configuration Management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tqdm>=4.66.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.84.0",
    
    # Code Quality
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    
    # Documentation
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "myst-parser>=2.0.0",
    
    # Debugging & Development
    "ipython>=8.17.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "notebook>=7.0.0",
    "ipdb>=0.13.0",
    "rich>=13.7.0",
    
    # Profiling & Performance
    "py-spy>=0.3.14",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/example/stock-predictor-prod"
Documentation = "https://stock-predictor-prod.readthedocs.io"
Repository = "https://github.com/example/stock-predictor-prod.git"
Issues = "https://github.com/example/stock-predictor-prod/issues"

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | data
  | notebooks
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*", "**/node_modules/*"]
known_first_party = ["production"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
exclude = ["tests/", "notebooks/"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --cov=production --cov-report=html --cov-report=term-missing"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "performance: Performance tests",
    "slow: Slow running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["production"]
omit = [
    "*/tests/*",
    "*/site-packages/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.pylint."messages control"]
disable = [
    "C0330", "C0326",  # Bad whitespace
    "R0913",           # Too many arguments
    "R0914",           # Too many local variables
    "W0212",           # Protected access
]

[tool.pydantic-settings]
env_file = ".env"
case_sensitive = false
